Split publish check integration tests
